Use an array of subdomains to refine mesh and optimization fixes
A different fix for a previous bug that fixed with another way. The bug was that when the indecesOfTrianglesThatNeedRefinement has only one element then the refinemesh function can not produce a new mesh and returns the same mesh, that has as a result the program to go in an infinte loop state because it calculates the same error again and again. Now we convert/match indecesOfTrianglesThatNeedRefinement to an array of subdomains in order to use this to refine mesh at each step. With that solution if the indecesOfTrianglesThatNeedRefinement  has only one element we can still refine the mesh.