Refactor of scripts to method, add a script and a fix
Combine solvePDEProblemWithAdaptiveFEM script and solvePDEProblemWithRegularFEM script to one method  - solvePDEProblemWithFEM which with different mode either solve the problem with Regular FEM or Adaptive FEM. Also add a script that runs in loop for different values the solvePDEProblemWithFEM and produce different figures with the result. At the end produce a log log plot. And finally add a fix.